Support mulitple Surfaces of react-art with each other mode in a document
react-art doesn't provide a way to use mutiple Surfaces with each other mode (e.g. svg, canvas, vml)
This can't cover cases which use one more Surfaces with each other mode in a document.

This commit enables react-art to provide the following features
 * To set Suface mode like <Surface mode='svg'> or <Surface mode='canvas'>
   If mode is omitted or not available value, canvas is used as default mode.

 * To use multiple Surface Components at the same time in same document,
   even if they set each other mode of surface

 * To set mode of Path utility like Path(path, 'svg') or Path(path, 'canvas').
   If Path() or Path(path) are used without mode, canvas is used as default mode.
   Path which react-art exports is just utility, not Component(e.g. Shape, Group) which
   can be handled selectively by mode in createInstance of ReactARTHostConfig.
   Thus this commit provides a way that users of react-art set mode of Path.

Some test cases are added for validating this commit, and
the following mock-up functions of DOM are also added to Jest environment for test.
 * window.SVGElement.prototype.getBBox
 * window.HTMLCanvasElement.prototype.getContext

The fixture regarding react-art is also updated to show this commit work well.